More robust method of initialising VB-GMM posteriors for CT
For CT, the VB-GMM posteriors and priors are now initialised in a hopefully more robust way. This is by dividing the accumulated histograms into three parts: background, brain and bone, and then fitting a VB-GMM to each partition. Ideally the lower bound of the fit should be used to do ARD for selecting the number of gaussians, but there seems to be an issue with the VB-GMM implementation in spm_imbasics at the moment, so it doesn't work. I will look into this.

If there are healthy subjects, one class in this population will be picked (the one closest to the lesion class in the diseased population) and set to zero.